requirements.txt
asgiref==3.7.2
Django==4.2.2
django-admin-interface==0.26.0
django-colorfield==0.9.0
django-cors-headers==4.1.0
django-filter==23.2
django-flat-responsive==2.0
djangorestframework==3.14.0
Pillow==9.5.0
python-decouple==3.8
python-slugify==8.0.1
pytz==2023.3
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
